+++ models/comment.js
@@ -2,6 +2,7 @@ var mongoose = require ("mongoose");
 
 var commentSchema = new mongoose.Schema({
     text: String,
+    updatedAt: { type: Date, default: Date.now },
     author: {
         type: mongoose.Schema.Types.ObjectId,
         ref: "User"