+++ models/user.js
@@ -5,7 +5,8 @@ var passportLocalMongoose = require("passport-local-mongoose");
 var userSchema = new mongoose.Schema ({
     username: { type: String, required: [true, "can't be blank"], unique: true, index: true},
     email: { type: String, required: [true, "can't be blank"], unique: true, index: true},
-    password: String
+    password: String,
+    isAdmin: {type: Boolean, default: false}
 }, {timestamps: true});
 
 userSchema.plugin(passportLocalMongoose);